chore: move withencoded struct
